refactor: message 상수를 관리하는 파일 이름 변경
